Move Python interface out of config loader
